requirements.txt
asgiref==3.2.3
certifi==2019.9.11
dj-database-url==0.5.0
Django==3.0.3
django-cors-headers==3.2.1
django-heroku==0.3.1
djangorestframework==3.11.0
gunicorn==20.0.4
pipenv==2018.11.26
psycopg2==2.8.4
pytz==2019.3
sqlparse==0.3.0
virtualenv==16.7.5
virtualenv-clone==0.5.3
virtualenvwrapper-win==1.2.5
whitenoise==5.0.1
